zgea/requirements.txt
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
idna==3.3
inflection==0.5.1
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.0.3
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==8.12.0
multitasking==0.0.10
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nbmake==1.3.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.22.3
numpy-financial==1.0.0
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
plotly==5.6.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.1.1
pytest-cov==6.3.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
Quandl==3.7.0
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
tenacity==8.0.1
terminado==0.13.3
testpath==0.6.0
tornado==6.1
tox==3.25.0
tqdm==4.64.0
traitlets==5.1.1
typeguard==2.13.3
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
yfinance==0.1.70